+++ src/main/java/module-info.java
@@ -18,9 +18,9 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-module jobj.reflect.main
+module io.github.astrapisixtynine.jobj.reflect
 {
-	requires lombok;
+	requires static lombok;
 	requires java.logging;
 	requires jobj.core.main;
 	requires modjenesis.main;